requirements.txt
yfinance
google-generativeai
tweepy
python-dotenv
requests
